refactor: consolidate storage management by introducing StorageProvider across desktop and extension clients, replacing WriteStorageProvider and updating related components for improved clarity and functionality
